more minor changes, may need double checking on final return in reg score
